moving script file
