+++ apps/web/public/static/locales/lv/common.json
@@ -125,5 +125,5 @@
   "create_account": "Izveidot Kontu",
   "confirm_password": "Apstiprināt paroli",
   "already_have_account": "Vai jums jau ir konts?",
-  "signing_up_terms": "Turpinot, jūs piekrītat <0>noteikumiem</0> un <1>privātuma politikai<1>",
+  "signing_up_terms": "Turpinot, jūs piekrītat <0>noteikumiem</0> un <1>privātuma politikai<1>"
 }
